requirements.txt
black
python-dotenv
aiogram==2.25.1